client - refactored all files named "index.js"
