+++ client/src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import {Provider} from "react-redux";
-import store from "./redux/store";
-import AppRouter from "./route";
+import store from "./redux/store/store";
+import AppRouter from "./route/AppRouter";
 
 ReactDOM.render(
     <Provider store={store}>
