+++ client/src/redux/middleware/Logger.js
@@ -1,5 +1,5 @@
 
-export const loggerTest = store => next => action => {
+export const logger = store => next => action => {
     console.log('[Log] Action\'s type is ', action);
     let result = next(action);
     console.log('[Log] Changed state is ', store.getState());
