--- client/src/redux/store/index.js
@@ -1,20 +1,0 @@
-import {createStore, applyMiddleware} from "redux";
-import thunk from "redux-thunk";
-
-import {rootReducer} from "../reducer";
-import {
-    loggerTest,
-} from "../middleware";
-
-const store = createStore(
-    rootReducer,
-    applyMiddleware(
-        thunk,
-        loggerTest,
-    )
-);
-
-// todo STOPSHIP remove , for testing purposes only
-window.store = store;
-
-export default store;
